refactor : FeedDetail 리팩토링
