+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/apiModel/response/FeedResponse.kt
@@ -1,0 +1,26 @@
+package com.emmsale.data.apiModel.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class FeedResponse(
+    @SerialName("id")
+    val id: Long,
+    @SerialName("eventId")
+    val eventId: Long,
+    @SerialName("title")
+    val title: String,
+    @SerialName("content")
+    val content: String,
+    @SerialName("writer")
+    val writer: MemberResponse,
+    @SerialName("images")
+    val imageUrls: List<String>,
+    @SerialName("commentCount")
+    val commentCount: Int,
+    @SerialName("createdAt")
+    val createdAt: String,
+    @SerialName("updatedAt")
+    val updatedAt: String,
+)
