--- android/2023-emmsale/app/src/main/java/com/emmsale/data/mapper/WriterMapper.kt
@@ -1,10 +1,0 @@
-package com.emmsale.data.mapper
-
-import com.emmsale.data.apiModel.response.WriterResponse
-import com.emmsale.data.model.Writer
-
-fun WriterResponse.toData(): Writer = Writer(
-    id = memberId,
-    name = name,
-    imageUrl = imageUrl,
-)
