+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/model/Feed.kt
@@ -2,17 +2,6 @@ package com.emmsale.data.model
 
 import java.time.LocalDateTime
 
-data class FeedDetail(
-    val id: Long,
-    val eventId: Long,
-    val title: String,
-    val content: String,
-    val writer: Writer,
-    val imageUrls: List<String>,
-    val createdAt: LocalDateTime,
-    val updatedAt: LocalDateTime,
-)
-
 data class Feed(
     val id: Long,
     val eventId: Long,
