--- android/2023-emmsale/app/src/main/java/com/emmsale/data/model/Writer.kt
@@ -1,7 +1,0 @@
-package com.emmsale.data.model
-
-data class Writer(
-    val id: Long,
-    val name: String,
-    val imageUrl: String,
-)
