+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/repository/concretes/DefaultFeedRepository.kt
@@ -1,11 +1,9 @@
 package com.emmsale.data.repository.concretes
 
-import com.emmsale.data.apiModel.response.FeedDetailResponse
 import com.emmsale.data.apiModel.response.FeedResponse
 import com.emmsale.data.common.retrofit.callAdapter.ApiResponse
 import com.emmsale.data.mapper.toData
 import com.emmsale.data.model.Feed
-import com.emmsale.data.model.FeedDetail
 import com.emmsale.data.repository.interfaces.FeedRepository
 import com.emmsale.data.service.FeedService
 import com.emmsale.di.modules.other.IoDispatcher
@@ -32,17 +30,9 @@ class DefaultFeedRepository @Inject constructor(
 
     override suspend fun getFeed(
         feedId: Long,
-    ): ApiResponse<FeedDetail> = withContext(dispatcher) {
-        feedService
-            .getFeed(feedId)
-            .map(FeedDetailResponse::toData)
-    }
-
-    override suspend fun getFeed2(
-        feedId: Long,
     ): ApiResponse<Feed> = withContext(dispatcher) {
         feedService
-            .getFeed2(feedId)
+            .getFeed(feedId)
             .map(FeedResponse::toData)
     }
