+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/feedDetail/uiState/FeedDetailUiState.kt
@@ -1,14 +1,14 @@
 package com.emmsale.presentation.ui.feedDetail.uiState
 
-import com.emmsale.data.model.FeedDetail
-import com.emmsale.data.model.Writer
+import com.emmsale.data.model.Feed
+import com.emmsale.data.model.Member
 import com.emmsale.presentation.common.FetchResult
 import com.emmsale.presentation.common.FetchResultUiState
 import java.time.LocalDateTime
 
 data class FeedDetailUiState(
     override val fetchResult: FetchResult,
-    val feedDetail: FeedDetail,
+    val feedDetail: Feed,
     val comments: List<CommentUiState>,
 ) : FetchResultUiState() {
 
@@ -27,17 +27,14 @@ data class FeedDetailUiState(
     companion object {
         val Loading: FeedDetailUiState = FeedDetailUiState(
             fetchResult = FetchResult.LOADING,
-            feedDetail = FeedDetail(
-                id = -1,
-                eventId = -1,
+            feedDetail = Feed(
+                id = 0,
+                eventId = 0,
                 title = "",
                 content = "",
-                writer = Writer(
-                    id = -1,
-                    name = "",
-                    imageUrl = "",
-                ),
+                writer = Member(),
                 imageUrls = emptyList(),
+                commentCount = 0,
                 createdAt = LocalDateTime.now(),
                 updatedAt = LocalDateTime.now(),
             ),
