+++ android/2023-emmsale/app/src/main/res/layout/item_feeddetail_feed_detail.xml
@@ -32,7 +32,7 @@
             android:layout_marginTop="14dp"
             android:contentDescription="@string/feeddetail_author_image_description"
             android:onClick="@{() -> onProfileImageClick.invoke(uiState.feedDetail.writer.id)}"
-            app:imageUrl="@{uiState.feedDetail.writer.imageUrl}"
+            app:imageUrl="@{uiState.feedDetail.writer.profileImageUrl}"
             app:isCircle="@{true}"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
